Extract region planning helpers
